+++ src/main/java/io/drakon/pulsar/internal/Configuration.java
@@ -8,7 +8,7 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
-import cpw.mods.fml.common.Loader;
+import net.minecraftforge.fml.common.Loader;
 import io.drakon.pulsar.config.IConfiguration;
 import io.drakon.pulsar.internal.logging.ILogger;
 import io.drakon.pulsar.pulse.PulseMeta;
@@ -50,7 +50,7 @@ public class Configuration implements IConfiguration {
     public boolean isModuleEnabled(PulseMeta meta) {
         ConfigEntry entry = modules.get(meta.getId());
         if (entry == null) {
-            modules.put(meta.getId(), new ConfigEntry(meta.isEnabled(), meta.getDescription()));
+            modules.put(meta.getId(), new ConfigEntry(meta.isDefaultEnabled(), meta.getDescription()));
             return meta.isEnabled();
         } else {
             return entry.getEnabled();