+++ test/io/drakon/pulsar/test_mod/PulsarWorkbench.java
@@ -1,10 +1,11 @@
 package io.drakon.pulsar.test_mod;
 
-import cpw.mods.fml.common.Mod;
-import cpw.mods.fml.common.Mod.EventHandler;
-import cpw.mods.fml.common.event.FMLInitializationEvent;
-import cpw.mods.fml.common.event.FMLPostInitializationEvent;
-import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import io.drakon.pulsar.debug.EventSpy;
+import io.drakon.pulsar.internal.logging.ILogger;
+import io.drakon.pulsar.internal.logging.LogManager;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.Mod.EventHandler;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import io.drakon.pulsar.control.PulseManager;
 
 /**
@@ -12,29 +13,26 @@ import io.drakon.pulsar.control.PulseManager;
  *
  * @author Arkan <arkan@drakon.io>
  */
+@SuppressWarnings("ALL")  // It's a dev env. Welcome to hacky hacky land.
 @Mod(modid = "PulsarWorkbench", name = "Pulsar Workbench", version = "DEV")
 public class PulsarWorkbench {
 
-    private PulseManager mgr;
+    private static PulseManager mgr;
+    private static ILogger log;
+
+    static {
+        mgr = new PulseManager("PulsarWorkbench");
+        log = LogManager.getLogger("PulsarWorkbench");
+
+        mgr.registerPulse(new EventSpy());
+
+        // Comment this out or it'll crash by design.
+        //mgr.registerPulse(new CrashyAsFuckPulse());
+    }
 
     @EventHandler
     public void preinit(FMLPreInitializationEvent evt) {
-        mgr = new PulseManager("PulsarWorkbench", "PulsarWorkbench");
-
-        // Comment this out or it'll crash by design.
-        mgr.registerPulse(new CrashyAsFuckPulse());
-
-        mgr.preInit(evt);
-    }
-
-    @EventHandler
-    public void init(FMLInitializationEvent evt) {
-        mgr.init(evt);
-    }
-
-    @EventHandler
-    public void postinit(FMLPostInitializationEvent evt) {
-        mgr.postInit(evt);
+        log.info("Workbench preinit.");
     }
 
 }